+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.0.2] - (TBC)
 
+### Feature
+
+- [07](http://tickets.projectname.com/browse/PROJECTNAME-RRRR)
+create form for creating recruitments
+
 ### Technical
 
 - [06](http://tickets.projectname.com/browse/PROJECTNAME-RRRR)